+++ library/src/EmitterConfig.ts
@@ -5,7 +5,7 @@ import { RotationBehaviorConfig } from "./behavior/built-in/RotationBehavior";
 import { ScaleBehaviorConfig } from "./behavior/built-in/ScaleBehavior";
 import { SpawnBehaviorConfig } from "./behavior/built-in/SpawnBehavior";
 import { TextureBehaviorConfig } from "./behavior/built-in/TextureBehavior";
-import { Ease } from "./data/Ease";
+import { Ease } from "./data/easing/Ease";
 
 /**
  * Type describing the configuration options for an Emitter.
